backend/requirements.txt
flask
pytest
psycopg2-binary
dotenv

